moved inspect_string_with_exclusion_dict_substring into separate file
