+++ dlp/snippets/custom_infotype.py
@@ -94,84 +94,6 @@ def inspect_string_with_exclusion_regex(
 # [END dlp_inspect_string_with_exclusion_regex]
 
 
-# [START dlp_inspect_string_with_exclusion_dict_substring]
-from typing import List  # noqa: F811, E402, I100
-
-import google.cloud.dlp  # noqa: F811, E402
-
-
-def inspect_string_with_exclusion_dict_substring(
-    project: str, content_string: str, exclusion_list: List[str] = ["TEST"]
-) -> None:
-    """Inspects the provided text, avoiding matches that contain excluded tokens
-
-    Uses the Data Loss Prevention API to omit matches if they include tokens
-    in the specified exclusion list.
-
-    Args:
-        project: The Google Cloud project id to use as a parent resource.
-        content_string: The string to inspect.
-        exclusion_list: The list of strings to ignore partial matches on
-
-    Returns:
-        None; the response from the API is printed to the terminal.
-    """
-
-    # Instantiate a client.
-    dlp = google.cloud.dlp_v2.DlpServiceClient()
-
-    # Construct a list of infoTypes for DLP to locate in `content_string`. See
-    # https://cloud.google.com/dlp/docs/concepts-infotypes for more information
-    # about supported infoTypes.
-    info_types_to_locate = [{"name": "EMAIL_ADDRESS"}, {"name": "DOMAIN_NAME"}]
-
-    # Construct a rule set that will only match if the match text does not
-    # contains tokens from the exclusion list.
-    rule_set = [
-        {
-            "info_types": info_types_to_locate,
-            "rules": [
-                {
-                    "exclusion_rule": {
-                        "dictionary": {"word_list": {"words": exclusion_list}},
-                        "matching_type": google.cloud.dlp_v2.MatchingType.MATCHING_TYPE_PARTIAL_MATCH,
-                    }
-                }
-            ],
-        }
-    ]
-
-    # Construct the configuration dictionary
-    inspect_config = {
-        "info_types": info_types_to_locate,
-        "rule_set": rule_set,
-        "include_quote": True,
-    }
-
-    # Construct the `item`.
-    item = {"value": content_string}
-
-    # Convert the project id into a full resource id.
-    parent = f"projects/{project}"
-
-    # Call the API.
-    response = dlp.inspect_content(
-        request={"parent": parent, "inspect_config": inspect_config, "item": item}
-    )
-
-    # Print out the results.
-    if response.result.findings:
-        for finding in response.result.findings:
-            print(f"Quote: {finding.quote}")
-            print(f"Info type: {finding.info_type.name}")
-            print(f"Likelihood: {finding.likelihood}")
-    else:
-        print("No findings.")
-
-
-# [END dlp_inspect_string_with_exclusion_dict_substring]
-
-
 # [START dlp_inspect_string_without_overlap]
 import google.cloud.dlp  # noqa: F811, E402
